add metodos item formula na biblioteca

+++ Biblioteca/Negocio/FormulaNegocio.cs
@@ -14,7 +14,7 @@ namespace Biblioteca.Negocio
         {
             if (produto == null)
             {
-                throw new Exception("Informar os dados do produto");
+                throw new Exception("Informar os dados da fórmula");
             }
 
             if (produto.Id <= 0)
@@ -24,7 +24,7 @@ namespace Biblioteca.Negocio
 
             if (produto.MateriaPrima == null)
             {
-                throw new Exception("Informar os dados da materia prima");
+                throw new Exception("Informar os dados da matéria prima");
             }
 
             FormulaDados dados = new FormulaDados();
@@ -40,8 +40,58 @@ namespace Biblioteca.Negocio
             {
                 if (dadosMateriaPrima.VerificarDuplicidade(materiaPrima) == false)
                 {
-                    throw new Exception("A materia prima não esta cadastrado");
+                    throw new Exception("A matéria prima não esta cadastrada");
                 }
+            }
+
+            if (this.VerificarDuplicidadeFormula(produto) == true)
+            {
+                throw new Exception("A fórmula já se encontra cadastrada");
+            }
+
+            dados.CadastrarFormula(produto);
+        }
+
+        public void CadastrarItemFormula(Produto produto)
+        {
+            if (produto == null)
+            {
+                throw new Exception("Informar os dados da fórmula");
+            }
+
+            if (produto.Id <= 0)
+            {
+                throw new Exception("O Id do produto não poderá ser menor ou igual a zero");
+            }
+
+            if (produto.MateriaPrima == null)
+            {
+                throw new Exception("Informar os dados da materia prima");
+            }
+
+            if (produto.MateriaPrima.First().Id <= 0)
+            {
+                throw new Exception("O Id da matéria prima não poderá ser menor ou igual a zero");
+            }
+
+            FormulaDados dados = new FormulaDados();
+            ProdutoNegocio dadosProduto = new ProdutoNegocio();
+            MateriaPrimaNegocio dadosMateriaPrima = new MateriaPrimaNegocio();
+
+            if (dadosProduto.VerificarDuplicidade(produto) == false)
+            {
+                throw new Exception("O produto não esta cadastrado");
+            }
+
+            
+            if (dadosMateriaPrima.VerificarDuplicidade(produto.MateriaPrima.First()) == false)
+            {
+                throw new Exception("A matéria prima não esta cadastrada");
+            }
+
+            if (this.VerificarDuplicidadeItemFormula(produto) == true)
+            {
+                throw new Exception("O item da fórmula já se encontra cadastrado");
             }
 
             dados.CadastrarFormula(produto);
@@ -61,7 +111,40 @@ namespace Biblioteca.Negocio
 
             if (this.VerificarDuplicidadeFormula(produto) == false)
             {
-                throw new Exception("A formula não esta cadastrado");
+                throw new Exception("A fórmula não esta cadastrada");
+            }
+
+            dados.DeleteFormula(produto);
+        }
+
+        public void DeleteItemFormula(Produto produto)
+        {
+            if (produto == null)
+            {
+                throw new Exception("Informar os dados da Fórmula");
+            }
+            if (produto.Id <= 0)
+            {
+                throw new Exception("O Id do produto não poderá ser menor ou igual a zero");
+            }
+            if (produto.MateriaPrima == null)
+            {
+                throw new Exception("Informar os dados da materia prima");
+            }
+            if (produto.MateriaPrima.First().Id <= 0)
+            {
+                throw new Exception("O Id da matéria prima não poderá ser menor ou igual a zero");
+            }
+            FormulaDados dados = new FormulaDados();
+
+            if (this.VerificarDuplicidadeFormula(produto) == false)
+            {
+                throw new Exception("A fórmula não esta cadastrada");
+            }
+
+            if (this.VerificarDuplicidadeItemFormula(produto) == false)
+            {
+                throw new Exception("O item da fórmula não esta cadastrado");
             }
 
             dados.DeleteFormula(produto);
@@ -79,6 +162,11 @@ namespace Biblioteca.Negocio
                 throw new Exception("O Id do produto não poderá ser menor ou igual a zero");
             }
             return new FormulaDados().SelectFormula(filtro);
+        }
+
+        public bool VerificarDuplicidadeItemFormula(Produto produto)
+        {
+            return new FormulaDados().VerificarDuplicidadeItemFormula(produto);
         }
 
         public bool VerificarDuplicidadeFormula(Produto produto)
